moving around firstrow
