+++ lib/Redesign/RecipePages/Recipe/BaseRecipe.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import 'first_row_widget.dart';
+import 'FirstRow/first_row_widget.dart';
 import 'second_row_widget.dart';
 import '../../components/ComponentsRedesign.dart';
