+++ lib/Redesign/RecipePages/Recipe/FirstRow/first_row_widget.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import '../../../components/ComponentsRedesign.dart';
+import '../../../components/StackedTiles.dart';
+import '../../../components/imagesclass.dart';
+import 'IntroductionBox.dart';
+import 'ImageBoxes.dart';
+
+class FirstRowWidget extends StatelessWidget {
+  const FirstRowWidget({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        const Widht40wallWidget(),
+        const IntroductionWidget(),
+        const Width20StandardWidget(),
+        MainImagePlusBoxes(),
+      ],
+    );
+  }
+}
